Responded to PR suggestions dquote> Moved BasePrivacyFilter to seperate mixin file; removed unused commented out code; removed redundant dependencies
