+++ analyses/models.py
@@ -20,6 +20,7 @@ from analyses.base_models.base_models import (
     ENADerivedManager,
     ENADerivedModel,
     MGnifyAutomatedModel,
+    PrivacyFilterManagerMixin,
     TimeStampedModel,
     VisibilityControlledModel,
 )
@@ -64,26 +65,6 @@ class Biome(TreeModel):
         return re.sub(r"[^a-zA-Z0-9._]", "", underscore_punctuated)
 
 
-class BasePrivacyFilterManager:
-    """
-    Base mixin providing common privacy filtering methods for studies
-    """
-
-    def get_queryset(self, include_private=False, private_only=False):
-        qs = super().get_queryset()
-        if private_only:
-            return qs.filter(is_private=True)
-        if not include_private:
-            return qs.filter(is_private=False)
-        return qs
-
-    def private_only(self):
-        """
-        Returns only private studies
-        """
-        return self.get_queryset(private_only=True)
-
-
 class StudyManager(models.Manager):
     async def get_or_create_for_ena_study(self, ena_study_accession):
         logging.info(f"Will get/create MGnify study for {ena_study_accession}")
@@ -106,7 +87,7 @@ class StudyManager(models.Manager):
         return study
 
 
-class PublicStudyManager(BasePrivacyFilterManager, StudyManager):
+class PublicStudyManagerMixin(PrivacyFilterManagerMixin, StudyManager):
     """
     A custom manager that filters out private studies by default.
     """
@@ -115,7 +96,7 @@ class PublicStudyManager(BasePrivacyFilterManager, StudyManager):
 
 
 class Study(MGnifyAutomatedModel, ENADerivedModel, TimeStampedModel):
-    objects = PublicStudyManager()
+    objects = PublicStudyManagerMixin()
     all_objects = models.Manager()
 
     accession = MGnifyAccessionField(
@@ -441,8 +422,8 @@ class AnalysisManagerIncludingAnnotations(models.Manager):
         return super().get_queryset()
 
 
-class PublicAnalysisManager(
-    BasePrivacyFilterManager, AnalysisManagerDeferringAnnotations
+class PublicAnalysisManagerMixin(
+    PrivacyFilterManagerMixin, AnalysisManagerDeferringAnnotations
 ):
     """
     A custom manager that filters out private analyses by default.
@@ -451,8 +432,8 @@ class PublicAnalysisManager(
     pass
 
 
-class PublicAnalysisManagerIncludingAnnotations(
-    BasePrivacyFilterManager, AnalysisManagerIncludingAnnotations
+class PublicAnalysisManagerMixinIncludingAnnotations(
+    PrivacyFilterManagerMixin, AnalysisManagerIncludingAnnotations
 ):
     """
     A custom manager that includes annotations but still filters out private analyses by default.
@@ -468,8 +449,8 @@ class Analysis(
     WithDownloadsModel,
     WithExperimentTypeModel,
 ):
-    objects = PublicAnalysisManager()
-    objects_and_annotations = PublicAnalysisManagerIncludingAnnotations()
+    objects = PublicAnalysisManagerMixin()
+    objects_and_annotations = PublicAnalysisManagerMixinIncludingAnnotations()
 
     all_objects = AnalysisManagerDeferringAnnotations()
     all_objects_and_annotations = AnalysisManagerIncludingAnnotations()
