+++ requirements.txt
@@ -32,7 +32,3 @@ requests~=2.32.3
 httpx~=0.27.2
 typing_extensions~=4.12.2
 asgiref~=3.8.1
-
-pytest~=8.3.4
-djangorestframework~=3.15.2
-responses~=0.25.3
